requirements.txt
requests
python-dotenv
pyaml-env
python-dxf